MD: Description Accuracy Disclaimer
This change adds a cautionary note about the accuracy of the
property descriptions. Since we don't have official docs at hand,
what we do is inferring what the meaning of each property is,
based on experiments we run.

That process, even though provided in good faith and with the
best effort, doesn't guarantee accuracy.