MD: Foundations for further Work (#58)
* MD: Description Accuracy Disclaimer

This change adds a cautionary note about the accuracy of the
property descriptions. Since we don't have official docs at hand,
what we do is inferring what the meaning of each property is,
based on experiments we run.

That process, even though provided in good faith and with the
best effort, doesn't guarantee accuracy.

* MD: Updated description for airResistanceFactor

Added one missing description as a test. The car data structure
goes much deeper and wider, therefore this table will probably
have to go away in any case. To be discussed.

* MD: High-Level Overview of Vehicle Performance

Added a high-level structure of the vehicle performance-related
parameters.

* MD: Addeed Chapters

Added chapters for the relevant data structures. This provides
the frame for future clarifications, as there are plenty of
things that need covering.

* MD: Added a TODO for Content Integration

The preexisting table now needs to be integrated into the above
structures. That may as well be a separate work item.