MD: Page Updates 2025-04-18 (#59)
* MD: Added vehAirControl

This change adds foundation for work on vehAirControl.

* MD: Expanded DriveModel with Motorcycle Info

Motorcycles have a few more variables, in addition to everthing
that the cars have. This change adds the foundation for the info
about those values.

* MD: Updated the Overall Data Structure

This is an attempt to rewrite the prior high-level structure
in the format that is overall present across this repo. It
will need to be revisited.

* MD: Correction of the Vehicle Name Prefix

In reality, there are multiple prefixes in use. Therefore, considering
just v_* isn't correct.

* MD: Importing and Adding Explanations of items

In this change:
  - A number of explanations was added to the page
  - The preexisting explanations were merged into the new
    information presentation

* MD: Added legend to the page's introduction

As to explain what certain nottations used throughout the page mean.