MD: Deleted Preexisting Content
As it has been fully replaced, in function if not in value,
by the new content.

The explanations from the preexisting content have been integrated
into the new content.

The 'default values' have been replaced with 'sample values' in
the new content. As far as I know, there is no such thing as
'default values' -- in principle, every vehicle has its own set
of values, characteristic for that vehicle. However, it is useful
for the reader to receive some reference values. Hence, preser-
vation of sample values as content.