MD: Added Sample Values & More (#60)
* MD: Moved the contribution history to the bottom

Its relevance is low to a random visitor who wants to learn
about vehciles. I'd rather defer to git history if there
was one, or automate this.

* MD: Added my Record to the History

Since we are having it.

* MD: Added Sample Values across all Tables

As far as the content goes, most of the variables have
no specific constraints. This lowers the content usefulness
to someone who experiments with their vehilce setup, as they
have no reference values on the page. Arguably, they would
have this in their own editor, but oooooooa page offering it
creates one more reference point to hang onto.

Hence, adding sample values. Most of the values come from
the in-game, non-modded, Caliburn.

* MD: Correction, ref -> TweakDBID

Instead of the term 'ref' for reference, let me use there the
term that seems common, for TweakDB IDs. Also, the sample values
are

  1) Making the sample column too wide
  2) Not as useful in a number
  3) Best expressed by the types in the high-level overview above

Hence, removing them, too.

* MD: Reworked the Legend

After all the changes, the legend could use an update to
account for all the content that is under.

* MD: Deleted Preexisting Content

As it has been fully replaced, in function if not in value,
by the new content.

The explanations from the preexisting content have been integrated
into the new content.

The 'default values' have been replaced with 'sample values' in
the new content. As far as I know, there is no such thing as
'default values' -- in principle, every vehicle has its own set
of values, characteristic for that vehicle. However, it is useful
for the reader to receive some reference values. Hence, preser-
vation of sample values as content.